Выносит функции по генерации случайных чисел в отдельный модуль, удаляет ненужные файлы
